Split docker workflow
Split docker workflow (build and push) into two jobs.

The first job will build the docker images and upload them as artifact. The second job will download them again and push them to the docker registry (if the build was started by a push on the main branch). This allows waiting for other jobs to be successful before pushing in the future.