Workaround broken `not-found` in next.js
`not-found.tsx` is prerendered for every page request, and as such calls the layouts `preload()` on every page, even if that specific layout is not used. So now we are skipping preload the hero layout is used in a not-found page. See vercel/next.js#61765

Additionaly the not-found page is not always rerendered when client navigating, and `headers()` gets outdated. Because of that I replaced the dynamic link to the legacy version with a static link to homepage :(