Allow job "Docker (arm64) / Build: web" to fail for PRs and merge queue
This job keeps failing because google fonts times out, so lets just allow this unless its a push to main. We catch all real errors in the amd64 build, which for some reason always works. We need the arm64 build to succeed on main, because thats the actual architecture of the production environment.