Fix new skins not connecting to achievement bits and items
If an achievement gets added that has `skin` bits or an item that unlocks a skin, but the skin does not exist in the db yet, once the skin gets added it would not connect to `achievementBits` or `unlockedByItems`. Ideally this should just run for new skins, but since there are missing connections already, we have to do this migration for all skins.